fix: use global config for ConfigProvider context
- Move global config functions to seperate file
- Reuse global config prefixCls for the context
- Move types to seperate file
